move Dashboard and Releases to screens
